Created seperate branches for future and listing gallery
